+++ services/rfq/relayer/inventory/export_test.go
@@ -9,3 +9,8 @@ import (
 func GetRebalance(cfg relconfig.Config, tokens map[int]map[common.Address]*TokenMetadata, chainID int, token common.Address) (*RebalanceData, error) {
 	return getRebalance(nil, cfg, tokens, chainID, token)
 }
+
+// GetRebalanceMetadatas is a wrapper around the internal getRebalanceMetadatas function.
+func GetRebalanceMetadatas(cfg relconfig.Config, tokens map[int]map[common.Address]*TokenMetadata, tokenName string, methods []relconfig.RebalanceMethod) (originTokenData, destTokenData *TokenMetadata, method relconfig.RebalanceMethod) {
+	return getRebalanceMetadatas(cfg, tokens, tokenName, methods)
+}